# license-application-service/src/main/java/de/hft/licensing/rest/PaymentsController.java
package de.hft.licensing.rest;

import de.hft.licensing.api.PaymentsApi;
import de.hft.licensing.db.enums.LicenseType;
import de.hft.licensing.db.enums.PaymentStatus;
import de.hft.licensing.db.tables.Application;
import de.hft.licensing.db.tables.ApplicationPayment;
import de.hft.licensing.db.tables.records.ApplicationPaymentRecord;
import de.hft.licensing.db.tables.records.ApplicationRecord;
import de.hft.licensing.model.ApplicationFeeResource;
import de.hft.licensing.model.ApplicationPaymentCreate;
import de.hft.licensing.model.ApplicationPaymentResource;
import de.hft.licensing.services.PaymentsControllerService;
import de.hft.licensing.services.dslService.BallotDslService;
import de.hft.licensing.utils.ApiFormValidator;
import de.hft.licensing.utils.RecordToResourceMapperUtil;
import org.jooq.DSLContext;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.http.ResponseEntity;
import org.springframework.security.access.prepost.PreAuthorize;
import org.springframework.transaction.annotation.Transactional;
import org.springframework.web.bind.annotation.RestController;

import java.math.BigDecimal;
import java.net.URI;
import java.time.Clock;
import java.time.LocalDateTime;
import java.util.List;

@RestController
public class PaymentsController implements PaymentsApi {

  private static final Logger log = LoggerFactory.getLogger(PaymentsController.class);
  private final int ETV_amount = 3500;
  private final int ETVPL_amount = 875;
  private final int ETV60_amount = 290;

  private final DSLContext dsl;
  private final BallotDslService ballotDslService;
  private final ApiFormValidator formValidator;
  private final PaymentsControllerService paymentsControllerService;

  public PaymentsController(DSLContext dsl, BallotDslService ballotDslService, ApiFormValidator formValidator, PaymentsControllerService paymentsControllerService) {
    this.dsl = dsl;
      this.ballotDslService = ballotDslService;
      this.formValidator = formValidator;
      this.paymentsControllerService = paymentsControllerService;
  }

  @Override
  @PreAuthorize("@paymentAuthorization.canAccessPayments(authentication, #applicationId)")
  @Transactional
  public ResponseEntity<ApplicationPaymentResource> createPayment(Integer applicationId,
      ApplicationPaymentCreate applicationPaymentCreate) {
    if (applicationId == null || applicationPaymentCreate.getApplicationId() == null) {
      log.error("Application ID is null");
      return ResponseEntity.badRequest().build();
    }
    if (applicationPaymentCreate.getName() != null &&
        applicationPaymentCreate.getIban() != null &&
        applicationPaymentCreate.getBic() != null &&
        !formValidator.isValidName(applicationPaymentCreate.getName()) &&
        !formValidator.isValidIban(applicationPaymentCreate.getIban()) &&
        !formValidator.isValidBic(applicationPaymentCreate.getBic())
    ) {
      log.error("Validation failed for payment creation");
      return ResponseEntity.badRequest().build();
    }
    LocalDateTime now = LocalDateTime.now(Clock.systemUTC());
    var dbPayment = dsl.insertInto(ApplicationPayment.APPLICATION_PAYMENT)
        .set(ApplicationPayment.APPLICATION_PAYMENT.PAYMENT_DATE, now)
        .set(ApplicationPayment.APPLICATION_PAYMENT.AMOUNT,paymentsControllerService.calculateFeeAmount(applicationId))
        .set(ApplicationPayment.APPLICATION_PAYMENT.ACCOUNTANT, applicationPaymentCreate.getName())
        .set(ApplicationPayment.APPLICATION_PAYMENT.IBAN, applicationPaymentCreate.getIban())
        .set(ApplicationPayment.APPLICATION_PAYMENT.BIC, applicationPaymentCreate.getBic())
        .set(ApplicationPayment.APPLICATION_PAYMENT.APPLICATION_ID, applicationId)
        .set(ApplicationPayment.APPLICATION_PAYMENT.PAYMENT_STATUS, PaymentStatus.unpaid)
        .set(ApplicationPayment.APPLICATION_PAYMENT.BALLOT_PERIOD_ID, ballotDslService.getCurrentBallotPeriodId())
        .returning()
        .fetchOneInto(ApplicationPaymentRecord.class);

    if (dbPayment == null) {
        log.error("Failed to create payment record for application ID {}", applicationId);
      return ResponseEntity.status(500).build();
    }
    log.info("Created payment record with ID {} for application ID {}", dbPayment.getId(), applicationId);

    ApplicationPaymentResource apiPayment = new ApplicationPaymentResource();
    RecordToResourceMapperUtil.mapApplicationPaymentRecordToResource(dbPayment, apiPayment);

    return ResponseEntity.created(
            URI.create("/applications/" + applicationId + "/payments/" + apiPayment.getId()))
        .body(apiPayment);
  }

  @Override
  @PreAuthorize("@paymentAuthorization.canAccessPayments(authentication, #applicationId)")
  public ResponseEntity<List<ApplicationPaymentResource>> listPayments(Integer applicationId) {
    if (applicationId == null) {
      return ResponseEntity.badRequest().build();
    }
    if (!dsl.fetchExists(
        dsl.selectOne()
            .from(Application.APPLICATION)
            .where(Application.APPLICATION.ID.eq(applicationId))
    )) {
      return ResponseEntity.notFound().build();
    }
    var payments = dsl.select()
        .from(ApplicationPayment.APPLICATION_PAYMENT)
        .where(ApplicationPayment.APPLICATION_PAYMENT.APPLICATION_ID.eq(applicationId))
        .fetchInto(ApplicationPaymentRecord.class);

    List<ApplicationPaymentResource> mappedPayments = payments.stream().map(record -> {
      ApplicationPaymentResource resource = new ApplicationPaymentResource();
      RecordToResourceMapperUtil.mapApplicationPaymentRecordToResource(record, resource);
      return resource;
    }).toList();

    return ResponseEntity.ok(mappedPayments);
  }

  @Override
  public ResponseEntity<ApplicationFeeResource> getApplicationFee(Integer applicationId) {

    if (applicationId == null) {
      return ResponseEntity.badRequest().build();
    }

    ApplicationRecord appRecord = dsl.select()
        .from(Application.APPLICATION)
        .where(Application.APPLICATION.ID.eq(applicationId))
        .fetchOneInto(ApplicationRecord.class);
    if (appRecord == null) {
      return ResponseEntity.notFound().build();
    }

    ApplicationFeeResource feeResource = new ApplicationFeeResource();
    feeResource.setApplicationId(applicationId);

    LicenseType applicationLicenseType = appRecord.getLicenseType();
    switch (applicationLicenseType) {
      case etv -> feeResource.setFeeAmount(new BigDecimal(ETV_amount));
      case etvpl -> feeResource.setFeeAmount(new BigDecimal(ETVPL_amount));
      case etv60 -> feeResource.setFeeAmount(new BigDecimal(ETV60_amount));
      default -> feeResource.setFeeAmount(BigDecimal.ZERO);
    }

    return ResponseEntity.ok(feeResource);
  }
}