# license-application-service/src/main/resources/api.yaml
openapi: 3.1.0
info:
  title: Mallorca Holiday Rental Licensing API
  version: 1.0.0
  description: |
    API for managing applications, licenses, and payments for the Mallorca Holiday Rental Licensing system.
    Supports GDPR-compliant user management through Keycloak integration and ensures transparent, fair allocation of ETV licenses.
servers:
  - url: https://localhost:8080/
...
security:
  - KeycloakOAuth2:
      - openid
      - profile
      - email
paths:
...
  /applications:
    get:
      tags: [Applications]
      summary: List all applications or filter by status
      description: Retrieve all holiday rental license applications or filter them by status and user ID.
      operationId: listApplications
      security:
        - KeycloakOAuth2: [openid]
      parameters:
        - name: user_id
          in: query
          description: Filter licenses by user ID
          required: false
          schema:
            type: string
            format: uuid
        - name: application_status
          in: query
          description: Filter applications by status
          required: false
          schema:
            $ref: "#/components/schemas/ApplicationStatusApiEnum"
      responses:
        "200":
          description: List of applications (optionally filtered by status)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApplicationResource"
    post:
      tags: [Applications]
      summary: Create a new application
      description: Submit a new holiday rental license application.
      operationId: createApplication
      security:
        - KeycloakOAuth2: [openid]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationCreate"
      responses:
        "201":
          description: Application created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationResource"
        "409":
          description: No active ballot period
        "422":
          description: User does not exist
...
  /applications/{applicationId}/payments:
    get:
      tags: [Payments]
      summary: List payments for an application
      description: Retrieve direct debit payment records associated with a specific application.
      operationId: listPayments
      security:
        - KeycloakOAuth2: [openid]
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payments for the given application
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApplicationPaymentResource"
    post:
      tags: [Payments]
      summary: Record a new payment for an application
      description: Record a new direct debit payment related to a holiday rental license application.
      operationId: createPayment
      security:
        - KeycloakOAuth2: [openid]
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationPaymentCreate"
      responses:
        "201":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationPaymentResource"
  /fees/{applicationId}:
    get:
      tags: [Payments]
      summary: Get application fee amount
      description: Retrieve the fee amount for a specific holiday rental license application.
      operationId: getApplicationFee
      security:
        - KeycloakOAuth2: [openid]
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application fee amount
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationFeeResource"
        "404":
          description: Application not found
...
components:
  securitySchemes:
    KeycloakOAuth2:
      type: oauth2
      description: OAuth2 with Keycloak JWT Bearer tokens.
      flows:
        authorizationCode:
          authorizationUrl: http://localhost:8080/auth/realms/license-realm/protocol/openid-connect/auth
          tokenUrl: http://localhost:8080/auth/realms/license-realm/protocol/openid-connect/token
          scopes:
            openid: OpenID Connect scope
            profile: Profile information access
            email: Email access
  schemas:
...
    ApplicationResource:
      type: object
      description: Represents a user's request for a holiday rental license.
      properties:
        id:
          type: integer
          description: Unique application identifier.
          example: 201
        user_id:
          type: string
          format: uuid
          description: ID of the applicant user.
          example: f28d1d3b-9bcb-4a74-a65a-2fede2b0a6c3
        license_type:
          $ref: "#/components/schemas/LicenseTypeApiEnum"
        cadastral_reference:
          type: string
          minLength: 20
          maxLength: 20
          description: Unique cadastral code (property reference).
          example: 1234567AB9999C0001DE
        applied_at:
          type: string
          format: date-time
          description: When the application was submitted.
          example: 2025-05-20T08:00:00Z
        changed_at:
          type: string
          format: date-time
          description: When the application was last updated.
          example: 2025-06-01T10:00:00Z
        application_status:
          $ref: "#/components/schemas/ApplicationStatusApiEnum"
        remarks:
          type: string
          description: Additional information or comments.
          example: Needs urgent review.
      required:
        - id
        - user_id
        - license_type
        - cadastral_reference
        - application_status
    ApplicationCreate:
      type: object
      description: Payload for creating a new application for a rental license.
      properties:
        user_id:
          type: string
          format: uuid
          description: User who submits the application.
          example: f28d1d3b-9bcb-4a74-a65a-2fede2b0a6c3
        license_type:
          $ref: "#/components/schemas/LicenseTypeApiEnum"
        cadastral_reference:
          type: string
          minLength: 20
          maxLength: 20
          description: Property cadastral reference.
          example: 1234567AB9999C0001DE
        remarks:
          type: string
          description: Additional notes.
          example: First time applying
      required:
        - user_id
        - license_type
        - cadastral_reference
...
    ApplicationPaymentResource:
      type: object
      description: Represents a payment made for an application.
      properties:
        id:
          type: integer
          description: Unique identifier for the payment.
          example: 401
        application_id:
          type: integer
          description: Associated application ID.
          example: 201
        amount:
          type: number
          format: decimal
          description: Payment amount in EUR.
          example: 123.45
        name:
          type: string
          description: Account holder's name.
          example: John Doe
        iban:
          type: string
          description: IBAN of the payment method.
          example: DE01234567890123456789
        bic:
          type: string
          description: BIC/SWIFT code of the bank.
          example: DEUTDEAAXXX
        payment_date:
          type: string
          format: date-time
          description: When payment was made.
          example: 2025-05-22T15:10:00Z
        payment_status:
          $ref: "#/components/schemas/PaymentStatusApiEnum"
      required:
        - id
        - application_id
        - amount
        - name
        - iban
        - payment_status
    ApplicationPaymentCreate:
      type: object
      description: Payload for creating a new payment for an application.
      properties:
        application_id:
          type: integer
          description: Associated application ID.
          example: 201
        name:
          type: string
          description: Account holder's name.
          example: John Doe
        iban:
          type: string
          description: IBAN of the payment method.
          example: DE01234567890123456789
        bic:
          type: string
          description: BIC/SWIFT code of the bank.
          example: DEUTDEAAXXX
      required:
        - application_id
        - name
        - iban
    ApplicationFeeResource:
      type: object
      description: Represents the fee amount for an application.
      properties:
        application_id:
          type: integer
          description: Associated application ID.
          example: 201
        fee_amount:
          type: number
          format: decimal
          description: Fee amount in EUR.
          example: 150.00
      required:
        - application_id
        - fee_amount
...
    ApplicationStatusApiEnum:
      type: string
      description: Status values for an application.
      enum:
        - DRAFT # Saved but not submitted
        - DOCUMENTS_SUBMITTED # Documents uploaded
        - VERIFICATION_PENDING # Awaiting verification
        - AWAITING_PAYMENT # Waiting for user payment
        - PAYMENT_RECEIVED # Payment details received
        - SUBMITTED # Application formally submitted
        - IN_BALLOT # Entered into selection/ballot
        - SELECTED # Chosen/approved via ballot
        - NOT_SELECTED # Not chosen in ballot
        - UNDER_REVIEW # In manual review process
        - APPROVED # License approved
        - REJECTED # Application rejected or payment denied
        - EXPIRED # Application expired (not continued)
        - CANCELLED # Withdrawn/cancelled by user
    PaymentStatusApiEnum:
      type: string
      description: Status for any payment attempt.
      enum:
        - UNPAID # Payment is pending
        - PAID # Payment completed
...
    LicenseTypeApiEnum:
      type: string
      description: License categories for holiday rentals.
      enum:
        - ETV
        - ETVPL
        - ETV60
...